Separate functions into new file
