+++ scorepostgenerator.py
@@ -1,7 +1,6 @@
 from ossapi import Ossapi
 from ossapi.enums import RankStatus
 from osu import Client
-from rosu_pp_py import Beatmap, Performance
 import dotenv
 import requests
 import os
@@ -9,6 +8,7 @@ from sys import exit
 import pyperclip as pc
 from circleguard import *
 import webbrowser
+from functions import calculate_pp, mod_sort
 
 dotenv.load_dotenv()
 
@@ -25,60 +25,6 @@ except PermissionError:
     quit()
 legacy_only = os.getenv("LEGACY_ONLY")
 api_osupy = Client.from_client_credentials(client_id, client_secret, callback_url)
-
-
-def acc_if_fc(score):
-    count300 = score.statistics.count_300
-    count100 = score.statistics.count_100
-    count50 = score.statistics.count_50
-    acc = (300 * count300 + 100 * count100 + 50 * count50) / (300 * (count300 + count100 + count50))
-    return acc * 100
-
-
-def downloadmap(score):
-    print("Downloading beatmap...")
-    url = "https://osu.ppy.sh/osu/" + str(score.beatmap.id)
-    r = requests.get(url)
-    open('osu.osu', 'wb').write(r.content)
-
-
-def calculatePP(function, score, maxcombo):
-    if maxcombo == 0: return 0
-    downloadmap(score)
-    mapfile = Beatmap(path="osu.osu")
-    calc = Performance(mods=score.mods.value)
-    if function == "curr":
-        print("Calculating score PP...")
-        calc.set_accuracy(score.accuracy * 100)
-        calc.set_misses(score.statistics.count_miss)
-        calc.set_combo(score.max_combo)
-    elif function == "fc":
-        print("Calculating PP if FC...")
-        calc.set_accuracy(acc_if_fc(score))
-        calc.set_misses(0)
-        calc.set_combo(maxcombo)
-    elif function == "ss":
-        print("Calculating PP if SS...")
-        calc.set_accuracy(100.00)
-        calc.set_misses(0)
-        calc.set_combo(maxcombo)
-    else:
-        print("Invalid entry.")
-        quit()
-    perf = calc.calculate(mapfile)
-    try:
-        os.remove("osu.osu")
-    except OSError as x:
-        print("Error occurred: %s : %s" % ("osu", x.strerror))
-    return perf.pp
-
-def mod_sort(mod):
-    mod_order = ['EZ', 'HD', 'DT', 'NC', 'HT', 'DC', 'HR', 'SD', 'PF', 'FL', 'RX', 'AP']
-    if mod in mod_order:
-        return mod_order.index(mod)
-    else:
-        return len(mod_order)
-
 
 gamemode = input("Mode (osu, taiko, fruits, mania): ")
 inputMode = input("User or score: ")
@@ -172,7 +118,7 @@ if beatmap.status == RankStatus.LOVED: post += "💖 "
 post += "| "
 
 if score.pp is None:
-    pp = calculatePP("curr", score, maxcombo)
+    pp = calculate_pp("curr", score, maxcombo)
 else:
     pp = score.pp
 if beatmap.status == RankStatus.RANKED:
@@ -183,7 +129,7 @@ if beatmap.status == RankStatus.RANKED:
 else:
     post += str(round(pp)) + "pp if ranked "
 if not score.perfect:
-    post += "(" + str(int(calculatePP("fc", score, maxcombo))) + "pp if FC) "
+    post += "(" + str(int(calculate_pp("fc", score, maxcombo))) + "pp if FC) "
 
 if score.replay:
     try:
